requirements.txt
faster-whisper
numpy
sounddevice
keyboard
pystray
Pillow
pyinstaller
huggingface_hub
